move ControllerMethodAnalyzer to Bridge/Symfony
