+++ src/Rector/Contrib/Symfony/Controller/ActionSuffixRemoverRector.php
@@ -3,7 +3,7 @@
 namespace Rector\Rector\Contrib\Symfony\Controller;
 
 use PhpParser\Node;
-use Rector\NodeAnalyzer\Contrib\Symfony\ControllerMethodAnalyzer;
+use Rector\Bridge\Symfony\NodeAnalyzer\ControllerMethodAnalyzer;
 use Rector\NodeChanger\IdentifierRenamer;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
