+++ src/Rector/Contrib/Symfony/HttpKernel/GetRequestRector.php
@@ -5,9 +5,9 @@ namespace Rector\Rector\Contrib\Symfony\HttpKernel;
 use PhpParser\Node;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Stmt\ClassMethod;
+use Rector\Bridge\Symfony\NodeAnalyzer\ControllerMethodAnalyzer;
 use Rector\Node\Attribute;
 use Rector\Node\NodeFactory;
-use Rector\NodeAnalyzer\Contrib\Symfony\ControllerMethodAnalyzer;
 use Rector\NodeAnalyzer\MethodCallAnalyzer;
 use Rector\NodeTraverserQueue\BetterNodeFinder;
 use Rector\Rector\AbstractRector;
